third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/edit/edit-style-attribute-expected.txt

Tests that style modification generates attribute updated event only when attribute is actually changed.


Running: testSetUp

Running: testSetNewValue
SDK.DOMModel.Events.AttrModified should be issued

Running: testSetSameValue
